requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
fonttools==4.59.1
fsspec==2025.7.0
gradio==5.43.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.12
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
orjson==3.11.2
packaging==25.0
pandas==2.3.2
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
roboflow==1.2.6
ruff==0.12.10
safehttpx==0.1.6
scipy==1.16.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
supervision==0.26.1
sympy==1.14.0
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.183
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
